Camera now works when the keyboard is opened.
In order to fix this I modified the libcamera to make trick the android
framework to think that there's 2 cameras in the backflip instead of
just one.

Users now have the possibility to change to a front facing camera when
they want to have proper preview and pictures when the keyboard is
opened.

Change-Id: Ib020a0387962b634dc26461d9071693a2b863691